Rename dashboard items to dashboard element
