+++ src/components/general/DashboardElement.vue
